v0.2: Windows GUI and genre resolver
Merge tested v0.2 GUI/resolver, CI fixes, media ignores, and validation baseline.

+++ src/genre_test/audio.py
@@ -27,7 +27,7 @@ def select_windows(
 ) -> list[np.ndarray]:
     if count < 1:
         raise ValueError("count must be >= 1")
-    width = max(1, int(round(window_seconds * sr)))
+    width = max(1, round(window_seconds * sr))
     n = audio.size
 
     if n <= width:
